+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 Added:
 - `.length` property to token, which returns the length of value matched
+- `empty()` top-level extension lexeme, same as `Lexeme.empty()`
 
 Changed:
 - Reorganized the documentation
 - Moved the utils directory inside the source directory
+- Renamed `grammar.lemexes` to `grammar.definitions`;
 
 ## 0.0.1